Modularize analytics pipeline (refactor)
Plus, fix PR completeness calc

- slice monolithic analytics aggregator into domain modules (events,
  reviews, repositories, profiles, engagement, individual) and import
  them from `analytics.ts`
- expose reusable Dependabot filter via shared helpers and write boundary
  tests for event/leaderboard helpers
- rebuild PR completeness leaderboard query to exclude self-reviews,
  pre-merge noise, and compute ratios; re-map UI layer to show per-PR
  averages
- add targeted Vitest coverage for response-time aggregation and
  DB-backed PR completeness scenario
