+++ src/lib/dashboard/analytics.ts
@@ -1,31 +1,64 @@
 import {
+  fetchActiveContributors,
+  fetchMainBranchContribution,
+  fetchReviewerActivity,
+  type ReviewerActivityRow,
+} from "@/lib/dashboard/analytics/engagement";
+import { fetchTotalEvents, fetchTrend } from "@/lib/dashboard/analytics/events";
+import {
+  fetchIndividualActivityHeatmap,
+  fetchIndividualCoverageMetrics,
+  fetchIndividualDiscussion,
+  fetchIndividualIssueMetrics,
+  fetchIndividualMergedByMetrics,
+  fetchIndividualMonthlyTrends,
+  fetchIndividualPrCompletionMetrics,
+  fetchIndividualPullRequestMetrics,
+  fetchIndividualRepoActivity,
+  fetchIndividualRepoComparison,
+  fetchIndividualReviewHeatmap,
+  fetchIndividualReviewMetrics,
+  type IndividualPrCompletionRow,
+} from "@/lib/dashboard/analytics/individual";
+import {
   buildMonthlyDurationTrend,
   fetchIssueAggregates,
   fetchIssueDurationDetails,
   summarizeIssueDurations,
 } from "@/lib/dashboard/analytics/issues";
+import {
+  fetchLeaderboard,
+  fetchPrCompletionLeaderboard,
+  type LeaderboardRow,
+  type PrCompletionLeaderboardRow,
+} from "@/lib/dashboard/analytics/leaderboards";
+import { resolveProfiles } from "@/lib/dashboard/analytics/profiles";
 import { fetchPrAggregates } from "@/lib/dashboard/analytics/pull-requests";
 import {
-  averageBusinessResponseHours,
+  fetchRepoComparison,
+  fetchRepoDistribution,
+  type RepoComparisonRawRow,
+  type RepoDistributionRow,
+} from "@/lib/dashboard/analytics/repositories";
+import {
+  fetchReviewAggregates,
+  fetchReviewHeatmap,
+} from "@/lib/dashboard/analytics/reviews";
+import {
   buildComparison,
   buildDurationComparison,
   buildHistorySeries,
   buildRatioComparison,
-  formatDateKey,
-  HOLIDAY_SET,
   normalizeText,
   resolveRange,
   roundToOneDecimal,
 } from "@/lib/dashboard/analytics/shared";
-import { calculateBusinessHoursBetween } from "@/lib/dashboard/business-days";
 import type {
   AnalyticsParams,
   DashboardAnalytics,
-  HeatmapCell,
   IndividualMetricHistory,
   LeaderboardEntry,
   LeaderboardSummary,
-  MultiTrendPoint,
   OrganizationAnalytics,
   RepoComparisonRow,
   RepoDistributionItem,
@@ -34,9 +67,7 @@ import type {
   WeekStart,
 } from "@/lib/dashboard/types";
 import { ensureSchema } from "@/lib/db";
-import { query } from "@/lib/db/client";
 import {
-  getRepositoryProfiles,
   getSyncConfig,
   getUserProfiles,
   listAllRepositories,
@@ -45,1966 +76,6 @@ import {
   type UserProfile,
 } from "@/lib/db/operations";
 import { env } from "@/lib/env";
-
-const DEPENDABOT_FILTER =
-  "NOT (COALESCE(LOWER(u.login), '') LIKE 'dependabot%' OR COALESCE(LOWER(u.login), '') = 'app/dependabot')";
-
-type ReviewAggregateRow = {
-  reviews_completed: number;
-  avg_response_hours: number | null;
-  avg_participation: number | null;
-};
-
-type ReviewStatsRow = {
-  reviews_completed: number;
-  avg_participation: number | null;
-};
-
-type ReviewResponseRow = {
-  reviewer_id: string | null;
-  pull_request_id: string;
-  requested_at: string;
-  responded_at: string | null;
-};
-
-async function fetchReviewAggregates(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<ReviewAggregateRow> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const [statsResult, responseRows] = await Promise.all([
-    query<ReviewStatsRow>(
-      `WITH pr_scope AS (
-         SELECT pr.id, pr.github_created_at, pr.github_merged_at, pr.repository_id, pr.author_id
-         FROM pull_requests pr
-         LEFT JOIN users u ON u.id = pr.author_id
-         WHERE pr.github_merged_at BETWEEN $1 AND $2${repoClause}
-           AND pr.github_merged_at IS NOT NULL
-           AND ${DEPENDABOT_FILTER}
-       ),
-       reviews_in_range AS (
-         SELECT r.id, r.author_id, r.github_submitted_at, r.pull_request_id
-         FROM reviews r
-         JOIN pr_scope pr ON pr.id = r.pull_request_id
-         WHERE r.github_submitted_at BETWEEN $1 AND $2
-       ),
-       participation AS (
-         SELECT
-           pr_scope.id AS pull_request_id,
-           COUNT(DISTINCT r.author_id)
-             FILTER (WHERE r.github_submitted_at BETWEEN $1 AND $2) AS reviewer_count
-         FROM pr_scope
-         LEFT JOIN reviews r ON r.pull_request_id = pr_scope.id
-         GROUP BY pr_scope.id
-       )
-       SELECT
-         (SELECT COUNT(*) FROM reviews_in_range) AS reviews_completed,
-         (SELECT AVG(COALESCE(participation.reviewer_count, 0)) FROM participation) AS avg_participation
-       `,
-      params,
-    ),
-    fetchReviewResponsePairs(start, end, repositoryIds),
-  ]);
-
-  const statsRow = statsResult.rows[0] ?? {
-    reviews_completed: 0,
-    avg_participation: null,
-  };
-
-  const avgResponseHours = averageBusinessResponseHours(
-    responseRows.map((row) => ({
-      requestedAt: row.requested_at,
-      respondedAt: row.responded_at,
-    })),
-    HOLIDAY_SET,
-  );
-
-  return {
-    reviews_completed: Number(statsRow.reviews_completed ?? 0),
-    avg_response_hours: avgResponseHours,
-    avg_participation: statsRow.avg_participation,
-  };
-}
-
-type TotalEventsRow = {
-  total_events: number;
-  issues: number;
-  pull_requests: number;
-  reviews: number;
-  comments: number;
-};
-
-async function fetchTotalEvents(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<TotalEventsRow> {
-  const params: unknown[] = [start, end];
-  let repoClauseIssues = "";
-  let repoClausePrs = "";
-  let repoClauseCommentsIssue = "";
-  let repoClauseCommentsPr = "";
-  let repoClauseReviews = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClauseIssues = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClausePrs = ` AND p.repository_id = ANY($${index}::text[])`;
-    repoClauseReviews = ` AND pr.repository_id = ANY($${index}::text[])`;
-    repoClauseCommentsIssue = ` AND ic.repository_id = ANY($${index}::text[])`;
-    repoClauseCommentsPr = ` AND pc.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<TotalEventsRow>(
-    `WITH issue_events AS (
-       SELECT COUNT(*) AS issues
-       FROM issues i
-       WHERE i.github_created_at BETWEEN $1 AND $2${repoClauseIssues}
-     ),
-     pr_events AS (
-       SELECT COUNT(*) AS pull_requests
-       FROM pull_requests p
-       LEFT JOIN users u ON u.id = p.author_id
-       WHERE p.github_created_at BETWEEN $1 AND $2${repoClausePrs}
-         AND ${DEPENDABOT_FILTER}
-     ),
-     review_events AS (
-       SELECT COUNT(*) AS reviews
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.github_submitted_at BETWEEN $1 AND $2${repoClauseReviews}
-         AND ${DEPENDABOT_FILTER}
-     ),
-     comment_events AS (
-       SELECT COUNT(*) AS comments
-       FROM comments c
-       LEFT JOIN LATERAL (
-         SELECT i.repository_id FROM issues i WHERE i.id = c.issue_id
-       ) ic ON TRUE
-       LEFT JOIN LATERAL (
-         SELECT p.repository_id FROM pull_requests p WHERE p.id = c.pull_request_id
-       ) pc ON TRUE
-       WHERE c.github_created_at BETWEEN $1 AND $2
-         AND (
-           (c.issue_id IS NOT NULL${repoClauseCommentsIssue})
-           OR (c.pull_request_id IS NOT NULL${repoClauseCommentsPr})
-         )
-     )
-     SELECT
-       issue_events.issues + pr_events.pull_requests + review_events.reviews + comment_events.comments AS total_events,
-       issue_events.issues,
-       pr_events.pull_requests,
-       review_events.reviews,
-       comment_events.comments
-     FROM issue_events, pr_events, review_events, comment_events`,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      total_events: 0,
-      issues: 0,
-      pull_requests: 0,
-      reviews: 0,
-      comments: 0,
-    }
-  );
-}
-
-type TrendRow = {
-  date: Date | string;
-  count: number | string;
-};
-
-async function fetchTrend(
-  table: "issues" | "pull_requests",
-  column: "github_created_at" | "github_closed_at" | "github_merged_at",
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  timeZone: string,
-): Promise<TrendPoint[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  const alias = table === "issues" ? "i" : "p";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND ${alias}.repository_id = ANY($${index}::text[])`;
-  }
-
-  const timezoneIndex = params.length + 1;
-  const queryText = `SELECT to_char(date_trunc('day', ${alias}.${column} AT TIME ZONE $${timezoneIndex}), 'YYYY-MM-DD') AS date, COUNT(*)
-    FROM ${table} ${alias}
-    WHERE ${alias}.${column} BETWEEN $1 AND $2${repoClause}
-    GROUP BY date
-    ORDER BY date`;
-
-  params.push(timeZone);
-  const result = await query<TrendRow>(queryText, params);
-  return result.rows.map((row) => {
-    let normalizedDate: string | null = null;
-    if (row.date instanceof Date) {
-      normalizedDate = formatDateKey(row.date);
-    } else if (typeof row.date === "string") {
-      normalizedDate = row.date;
-    }
-
-    return {
-      date: normalizedDate ?? String(row.date),
-      value: Number(row.count ?? 0),
-    };
-  });
-}
-
-type HeatmapRow = {
-  dow: number;
-  hour: number;
-  count: number;
-};
-
-async function fetchReviewHeatmap(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  timeZone: string,
-): Promise<HeatmapCell[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const timezoneIndex = params.length + 1;
-  const result = await query<HeatmapRow>(
-    `SELECT
-       EXTRACT(DOW FROM (r.github_submitted_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-       EXTRACT(HOUR FROM (r.github_submitted_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-       COUNT(*)
-     FROM reviews r
-     JOIN pull_requests pr ON pr.id = r.pull_request_id
-     WHERE r.github_submitted_at BETWEEN $1 AND $2${repoClause}
-      GROUP BY dow, hour
-      ORDER BY dow, hour`,
-    [...params, timeZone],
-  );
-
-  return result.rows.map((row) => ({
-    day: row.dow,
-    hour: row.hour,
-    count: Number(row.count ?? 0),
-  }));
-}
-
-async function fetchIndividualReviewHeatmap(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  timeZone: string,
-): Promise<HeatmapCell[]> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const timezoneIndex = params.length + 1;
-  const result = await query<HeatmapRow>(
-    `SELECT
-       EXTRACT(DOW FROM (r.github_submitted_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-       EXTRACT(HOUR FROM (r.github_submitted_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-       COUNT(*)
-     FROM reviews r
-     JOIN pull_requests pr ON pr.id = r.pull_request_id
-     LEFT JOIN users u ON u.id = pr.author_id
-     WHERE r.author_id = $1
-       AND r.github_submitted_at BETWEEN $2 AND $3${repoClause}
-       AND pr.author_id <> $1
-       AND COALESCE(r.state, '') <> 'DISMISSED'
-       AND ${DEPENDABOT_FILTER}
-     GROUP BY dow, hour
-     ORDER BY dow, hour`,
-    [...params, timeZone],
-  );
-
-  return result.rows.map((row) => ({
-    day: row.dow,
-    hour: row.hour,
-    count: Number(row.count ?? 0),
-  }));
-}
-
-async function fetchIndividualActivityHeatmap(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  timeZone: string,
-): Promise<HeatmapCell[]> {
-  const params: unknown[] = [personId, start, end];
-  let repoClauseIssues = "";
-  let repoClausePrAuthored = "";
-  let repoClausePrMerged = "";
-  let repoClauseReview = "";
-  let repoClauseCommentIssue = "";
-  let repoClauseCommentPr = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClauseIssues = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClausePrAuthored = ` AND p.repository_id = ANY($${index}::text[])`;
-    repoClausePrMerged = ` AND pr.repository_id = ANY($${index}::text[])`;
-    repoClauseReview = ` AND pr.repository_id = ANY($${index}::text[])`;
-    repoClauseCommentIssue = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClauseCommentPr = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const timezoneIndex = params.length + 1;
-  const result = await query<HeatmapRow>(
-    `WITH issue_events AS (
-       SELECT
-         EXTRACT(DOW FROM (i.github_created_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-         EXTRACT(HOUR FROM (i.github_created_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-         COUNT(*) AS count
-       FROM issues i
-       WHERE i.author_id = $1
-         AND i.github_created_at BETWEEN $2 AND $3${repoClauseIssues}
-       GROUP BY dow, hour
-     ),
-     pr_created_events AS (
-       SELECT
-         EXTRACT(DOW FROM (p.github_created_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-         EXTRACT(HOUR FROM (p.github_created_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-         COUNT(*) AS count
-       FROM pull_requests p
-       LEFT JOIN users u ON u.id = p.author_id
-       WHERE p.author_id = $1
-         AND p.github_created_at BETWEEN $2 AND $3${repoClausePrAuthored}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY dow, hour
-     ),
-     pr_merged_events AS (
-       SELECT
-         EXTRACT(DOW FROM (pr.github_merged_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-         EXTRACT(HOUR FROM (pr.github_merged_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-         COUNT(*) AS count
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.data -> 'mergedBy' ->> 'id' = $1
-         AND pr.github_merged_at IS NOT NULL
-         AND pr.github_merged_at BETWEEN $2 AND $3${repoClausePrMerged}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY dow, hour
-     ),
-     review_events AS (
-       SELECT
-         EXTRACT(DOW FROM (r.github_submitted_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-         EXTRACT(HOUR FROM (r.github_submitted_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-         COUNT(*) AS count
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1
-         AND r.github_submitted_at BETWEEN $2 AND $3${repoClauseReview}
-         AND pr.author_id <> $1
-         AND COALESCE(r.state, '') <> 'DISMISSED'
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY dow, hour
-     ),
-     comment_events AS (
-       SELECT
-         EXTRACT(DOW FROM (c.github_created_at AT TIME ZONE $${timezoneIndex}))::int AS dow,
-         EXTRACT(HOUR FROM (c.github_created_at AT TIME ZONE $${timezoneIndex}))::int AS hour,
-         COUNT(*) AS count
-       FROM comments c
-       LEFT JOIN issues i ON i.id = c.issue_id
-       LEFT JOIN pull_requests pr ON pr.id = c.pull_request_id
-       WHERE c.author_id = $1
-         AND c.github_created_at BETWEEN $2 AND $3
-         AND (
-           (c.issue_id IS NOT NULL${repoClauseCommentIssue})
-           OR (c.pull_request_id IS NOT NULL${repoClauseCommentPr})
-         )
-       GROUP BY dow, hour
-     ),
-     combined AS (
-       SELECT * FROM issue_events
-       UNION ALL
-       SELECT * FROM pr_created_events
-       UNION ALL
-       SELECT * FROM pr_merged_events
-       UNION ALL
-       SELECT * FROM review_events
-       UNION ALL
-       SELECT * FROM comment_events
-     )
-     SELECT dow, hour, SUM(count) AS count
-     FROM combined
-     GROUP BY dow, hour
-     ORDER BY dow, hour`,
-    [...params, timeZone],
-  );
-
-  return result.rows.map((row) => ({
-    day: row.dow,
-    hour: row.hour,
-    count: Number(row.count ?? 0),
-  }));
-}
-
-type RepoDistributionRow = {
-  repository_id: string;
-  issues: number;
-  pull_requests: number;
-  reviews: number;
-  comments: number;
-  total_events: number;
-};
-
-async function fetchRepoDistribution(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<RepoDistributionRow[]> {
-  const params: unknown[] = [start, end];
-  let repoFilter = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoFilter = ` AND repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<RepoDistributionRow>(
-    `WITH issue_counts AS (
-       SELECT repository_id AS repo_id, COUNT(*) AS issues
-       FROM issues
-       WHERE github_created_at BETWEEN $1 AND $2
-       GROUP BY repository_id
-     ),
-     pr_counts AS (
-       SELECT p.repository_id AS repo_id, COUNT(*) AS pull_requests
-       FROM pull_requests p
-       LEFT JOIN users u ON u.id = p.author_id
-       WHERE p.github_created_at BETWEEN $1 AND $2
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY p.repository_id
-     ),
-    review_counts AS (
-      SELECT pr.repository_id AS repo_id, COUNT(*) AS reviews
-      FROM reviews r
-      JOIN pull_requests pr ON pr.id = r.pull_request_id
-      LEFT JOIN users u ON u.id = pr.author_id
-      WHERE r.github_submitted_at BETWEEN $1 AND $2
-        AND ${DEPENDABOT_FILTER}
-      GROUP BY pr.repository_id
-    ),
-     comment_counts AS (
-       SELECT repository_id AS repo_id, COUNT(*) AS comments
-       FROM (
-         SELECT i.repository_id AS repository_id
-         FROM comments c
-         JOIN issues i ON i.id = c.issue_id
-         WHERE c.github_created_at BETWEEN $1 AND $2
-         UNION ALL
-         SELECT p.repository_id AS repository_id
-         FROM comments c
-         JOIN pull_requests p ON p.id = c.pull_request_id
-         LEFT JOIN users u ON u.id = p.author_id
-         WHERE c.github_created_at BETWEEN $1 AND $2
-           AND ${DEPENDABOT_FILTER}
-       ) AS combined
-       GROUP BY repository_id
-     ),
-    combined AS (
-      SELECT
-        COALESCE(ic.repo_id, pc.repo_id, rc.repo_id, cc.repo_id) AS repository_id,
-        COALESCE(ic.issues, 0) AS issues,
-        COALESCE(pc.pull_requests, 0) AS pull_requests,
-        COALESCE(rc.reviews, 0) AS reviews,
-        COALESCE(cc.comments, 0) AS comments
-      FROM issue_counts ic
-      FULL OUTER JOIN pr_counts pc ON pc.repo_id = ic.repo_id
-      FULL OUTER JOIN review_counts rc ON rc.repo_id = COALESCE(ic.repo_id, pc.repo_id)
-      FULL OUTER JOIN comment_counts cc ON cc.repo_id = COALESCE(ic.repo_id, pc.repo_id, rc.repo_id)
-    )
-     SELECT
-       repository_id,
-       issues,
-       pull_requests,
-      reviews,
-      comments,
-       (issues + pull_requests + reviews + comments) AS total_events
-     FROM combined
-     WHERE repository_id IS NOT NULL${repoFilter}
-     ORDER BY total_events DESC`,
-    params,
-  );
-
-  return result.rows;
-}
-
-type RepoComparisonRawRow = {
-  repository_id: string;
-  issues_created: number;
-  issues_resolved: number;
-  prs_created: number;
-  prs_merged: number;
-  prs_merged_by: number;
-  reviews: number;
-  active_reviews: number;
-  comments: number;
-  avg_first_review_hours: number | string | null;
-};
-
-async function fetchRepoComparison(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<RepoComparisonRawRow[]> {
-  const params: unknown[] = [start, end];
-  let repoFilterPr = "";
-  let repoFilterIssues = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoFilterPr = ` AND repository_id = ANY($${index}::text[])`;
-    repoFilterIssues = ` AND repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<RepoComparisonRawRow>(
-    `WITH repo_ids AS (
-       SELECT DISTINCT repository_id
-       FROM issues
-       WHERE github_closed_at BETWEEN $1 AND $2${repoFilterIssues}
-       UNION
-       SELECT DISTINCT repository_id
-       FROM issues
-       WHERE github_created_at BETWEEN $1 AND $2${repoFilterIssues}
-       UNION
-       SELECT DISTINCT pr.repository_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_merged_at BETWEEN $1 AND $2${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       UNION
-       SELECT DISTINCT pr.repository_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_created_at BETWEEN $1 AND $2${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       UNION
-       SELECT DISTINCT pr.repository_id
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.github_submitted_at BETWEEN $1 AND $2${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       UNION
-       SELECT DISTINCT repository_id
-       FROM (
-         SELECT i.repository_id AS repository_id
-         FROM comments c
-         JOIN issues i ON i.id = c.issue_id
-         WHERE c.github_created_at BETWEEN $1 AND $2${repoFilterIssues}
-         UNION
-         SELECT p.repository_id AS repository_id
-         FROM comments c
-         JOIN pull_requests p ON p.id = c.pull_request_id
-         LEFT JOIN users u ON u.id = p.author_id
-         WHERE c.github_created_at BETWEEN $1 AND $2${repoFilterPr}
-           AND ${DEPENDABOT_FILTER}
-       ) AS comment_repo_ids
-     ),
-     issue_created_counts AS (
-       SELECT repository_id, COUNT(*) AS issues_created
-       FROM issues
-       WHERE github_created_at BETWEEN $1 AND $2${repoFilterIssues}
-       GROUP BY repository_id
-     ),
-     issue_counts AS (
-       SELECT repository_id, COUNT(*) AS issues_resolved
-       FROM issues
-       WHERE github_closed_at BETWEEN $1 AND $2${repoFilterIssues}
-       GROUP BY repository_id
-     ),
-     pr_created_counts AS (
-       SELECT pr.repository_id, COUNT(*) AS prs_created
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_created_at BETWEEN $1 AND $2${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id
-     ),
-    pr_counts AS (
-      SELECT pr.repository_id, COUNT(*) AS prs_merged
-      FROM pull_requests pr
-      LEFT JOIN users u ON u.id = pr.author_id
-      WHERE pr.github_merged_at BETWEEN $1 AND $2${repoFilterPr}
-        AND ${DEPENDABOT_FILTER}
-      GROUP BY pr.repository_id
-    ),
-    pr_merged_by_counts AS (
-      SELECT pr.repository_id, COUNT(*) AS prs_merged_by
-      FROM pull_requests pr
-      LEFT JOIN users u ON u.id = pr.data -> 'mergedBy' ->> 'id'
-      WHERE pr.github_merged_at BETWEEN $1 AND $2${repoFilterPr}
-        AND pr.github_merged_at IS NOT NULL
-        AND ${DEPENDABOT_FILTER}
-      GROUP BY pr.repository_id
-    ),
-    review_counts AS (
-      SELECT
-        pr.repository_id,
-        COUNT(*) AS reviews,
-        COUNT(*) FILTER (WHERE r.state = 'APPROVED') AS active_reviews
-      FROM reviews r
-      JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.github_submitted_at BETWEEN $1 AND $2${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id
-     ),
-     comment_counts AS (
-       SELECT repository_id, COUNT(*) AS comments
-       FROM (
-         SELECT i.repository_id AS repository_id
-         FROM comments c
-         JOIN issues i ON i.id = c.issue_id
-         WHERE c.github_created_at BETWEEN $1 AND $2${repoFilterIssues}
-         UNION ALL
-         SELECT p.repository_id AS repository_id
-         FROM comments c
-         JOIN pull_requests p ON p.id = c.pull_request_id
-         LEFT JOIN users u ON u.id = p.author_id
-         WHERE c.github_created_at BETWEEN $1 AND $2${repoFilterPr}
-           AND ${DEPENDABOT_FILTER}
-       ) AS combined
-       GROUP BY repository_id
-     ),
-     first_review_times AS (
-       SELECT
-         pr.repository_id,
-         pr.github_created_at,
-         MIN(r.github_submitted_at) AS first_review_at
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       LEFT JOIN reviews r ON r.pull_request_id = pr.id
-       WHERE pr.github_created_at BETWEEN $1 AND $2${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id, pr.id, pr.github_created_at
-     ),
-     first_reviews AS (
-       SELECT
-         repository_id,
-         AVG(EXTRACT(EPOCH FROM (first_review_at - github_created_at)) / 3600.0) AS avg_first_review_hours
-       FROM first_review_times
-       WHERE first_review_at IS NOT NULL
-       GROUP BY repository_id
-     )
-     SELECT
-       repo_ids.repository_id,
-       COALESCE(issue_created_counts.issues_created, 0) AS issues_created,
-      COALESCE(issue_counts.issues_resolved, 0) AS issues_resolved,
-      COALESCE(pr_created_counts.prs_created, 0) AS prs_created,
-      COALESCE(pr_counts.prs_merged, 0) AS prs_merged,
-      COALESCE(pr_merged_by_counts.prs_merged_by, 0) AS prs_merged_by,
-      COALESCE(review_counts.reviews, 0) AS reviews,
-      COALESCE(review_counts.active_reviews, 0) AS active_reviews,
-      COALESCE(comment_counts.comments, 0) AS comments,
-      first_reviews.avg_first_review_hours
-    FROM repo_ids
-     LEFT JOIN issue_created_counts ON issue_created_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN issue_counts ON issue_counts.repository_id = repo_ids.repository_id
-      LEFT JOIN pr_created_counts ON pr_created_counts.repository_id = repo_ids.repository_id
-      LEFT JOIN pr_counts ON pr_counts.repository_id = repo_ids.repository_id
-      LEFT JOIN pr_merged_by_counts ON pr_merged_by_counts.repository_id = repo_ids.repository_id
-      LEFT JOIN review_counts ON review_counts.repository_id = repo_ids.repository_id
-      LEFT JOIN comment_counts ON comment_counts.repository_id = repo_ids.repository_id
-      LEFT JOIN first_reviews ON first_reviews.repository_id = repo_ids.repository_id`,
-    params,
-  );
-
-  return result.rows;
-}
-
-type ReviewerActivityRow = {
-  reviewer_id: string;
-  review_count: number;
-  prs_reviewed: number;
-  active_review_count: number;
-};
-
-type MainBranchContributionRow = {
-  user_id: string;
-  review_count: number;
-  active_review_count: number;
-  author_count: number;
-  additions: number;
-  deletions: number;
-  active_additions: number;
-  active_deletions: number;
-};
-
-async function fetchReviewerActivity(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  limit?: number,
-): Promise<ReviewerActivityRow[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  let limitClause = "";
-  if (typeof limit === "number" && Number.isFinite(limit) && limit > 0) {
-    params.push(limit);
-    limitClause = ` LIMIT $${params.length}`;
-  }
-
-  const result = await query<ReviewerActivityRow>(
-    `SELECT
-       r.author_id AS reviewer_id,
-       COUNT(*) FILTER (WHERE r.github_submitted_at BETWEEN $1 AND $2) AS review_count,
-       COUNT(DISTINCT r.pull_request_id) FILTER (WHERE r.github_submitted_at BETWEEN $1 AND $2) AS prs_reviewed,
-       COUNT(DISTINCT r.pull_request_id) FILTER (
-        WHERE r.github_submitted_at BETWEEN $1 AND $2 AND r.state = 'APPROVED'
-      ) AS active_review_count
-     FROM reviews r
-     JOIN pull_requests pr ON pr.id = r.pull_request_id
-     LEFT JOIN users u ON u.id = pr.author_id
-     WHERE r.author_id IS NOT NULL${repoClause}
-       AND COALESCE(r.state, '') <> 'DISMISSED'
-       AND ${DEPENDABOT_FILTER}
-     GROUP BY r.author_id
-     ORDER BY review_count DESC, prs_reviewed DESC${limitClause}`,
-    params,
-  );
-
-  return result.rows;
-}
-
-async function fetchMainBranchContribution(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<MainBranchContributionRow[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND p.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<MainBranchContributionRow>(
-    `WITH merged_prs AS (
-       SELECT
-         p.id,
-         p.author_id,
-         COALESCE((p.data ->> 'additions')::numeric, 0) AS additions,
-         COALESCE((p.data ->> 'deletions')::numeric, 0) AS deletions
-       FROM pull_requests p
-       LEFT JOIN users u ON u.id = p.author_id
-       WHERE p.github_merged_at BETWEEN $1 AND $2
-         AND p.github_merged_at IS NOT NULL${repoClause}
-         AND ${DEPENDABOT_FILTER}
-     ),
-     author_contrib AS (
-       SELECT
-         mp.author_id AS user_id,
-         COUNT(*) AS author_count,
-         SUM(mp.additions) AS additions,
-         SUM(mp.deletions) AS deletions
-       FROM merged_prs mp
-       WHERE mp.author_id IS NOT NULL
-       GROUP BY mp.author_id
-     ),
-    review_prs AS (
-      SELECT
-        r.author_id AS reviewer_id,
-        mp.id,
-        mp.additions,
-        mp.deletions,
-        BOOL_OR(r.state = 'APPROVED') AS has_approved
-      FROM merged_prs mp
-      JOIN reviews r ON r.pull_request_id = mp.id
-      LEFT JOIN users u ON u.id = r.author_id
-      WHERE r.author_id IS NOT NULL
-        AND ${DEPENDABOT_FILTER}
-        AND r.github_submitted_at BETWEEN $1 AND $2
-        AND COALESCE(r.state, '') <> 'DISMISSED'
-        AND r.author_id <> mp.author_id
-      GROUP BY r.author_id, mp.id, mp.additions, mp.deletions
-    ),
-    review_contrib AS (
-      SELECT
-        reviewer_id AS user_id,
-        COUNT(*) AS review_count,
-        SUM(CASE WHEN has_approved THEN 1 ELSE 0 END) AS active_review_count,
-        SUM(additions) AS additions,
-        SUM(deletions) AS deletions,
-        SUM(CASE WHEN has_approved THEN additions ELSE 0 END) AS active_additions,
-        SUM(CASE WHEN has_approved THEN deletions ELSE 0 END) AS active_deletions
-      FROM review_prs
-      GROUP BY reviewer_id
-    )
-    SELECT
-      COALESCE(author_contrib.user_id, review_contrib.user_id) AS user_id,
-      COALESCE(review_contrib.review_count, 0) AS review_count,
-      COALESCE(review_contrib.active_review_count, 0) AS active_review_count,
-      COALESCE(author_contrib.author_count, 0) AS author_count,
-      COALESCE(author_contrib.additions, 0) + COALESCE(review_contrib.additions, 0) AS additions,
-      COALESCE(author_contrib.deletions, 0) + COALESCE(review_contrib.deletions, 0) AS deletions,
-      COALESCE(author_contrib.additions, 0) + COALESCE(review_contrib.active_additions, 0) AS active_additions,
-      COALESCE(author_contrib.deletions, 0) + COALESCE(review_contrib.active_deletions, 0) AS active_deletions
-    FROM author_contrib
-    FULL OUTER JOIN review_contrib ON author_contrib.user_id = review_contrib.user_id
-    WHERE COALESCE(author_contrib.author_count, 0) + COALESCE(review_contrib.review_count, 0) > 0
-    ORDER BY (COALESCE(author_contrib.author_count, 0) + COALESCE(review_contrib.review_count, 0)) DESC,
-             COALESCE(author_contrib.additions, 0) + COALESCE(review_contrib.additions, 0) DESC`,
-    params,
-  );
-
-  return result.rows;
-}
-
-async function fetchReviewResponsePairs(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  reviewerId?: string,
-): Promise<ReviewResponseRow[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  let reviewerClause = "";
-  if (reviewerId) {
-    params.push(reviewerId);
-    reviewerClause = ` AND rr.reviewer_id = $${params.length}`;
-  }
-
-  const result = await query<ReviewResponseRow>(
-    `WITH pr_scope AS (
-       SELECT pr.id, pr.github_created_at, pr.repository_id, pr.author_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_created_at BETWEEN $1 AND $2${repoClause}
-         AND ${DEPENDABOT_FILTER}
-     ),
-     review_requests_scope AS (
-       SELECT rr.id, rr.pull_request_id, rr.reviewer_id, rr.requested_at, rr.removed_at
-       FROM review_requests rr
-       JOIN pr_scope pr ON pr.id = rr.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE rr.reviewer_id IS NOT NULL
-         AND rr.requested_at BETWEEN $1 AND $2${reviewerClause}
-         AND pr.author_id <> rr.reviewer_id
-         AND ${DEPENDABOT_FILTER}
-     ),
-     response_events AS (
-       SELECT
-         rr.id AS review_request_id,
-         r.github_submitted_at AS responded_at
-       FROM review_requests_scope rr
-       JOIN reviews r ON r.pull_request_id = rr.pull_request_id
-       WHERE r.author_id = rr.reviewer_id
-         AND r.github_submitted_at BETWEEN $1 AND $2
-       UNION ALL
-       SELECT
-         rr.id AS review_request_id,
-         c.github_created_at AS responded_at
-       FROM review_requests_scope rr
-       JOIN comments c ON c.pull_request_id = rr.pull_request_id
-       WHERE c.author_id = rr.reviewer_id
-         AND c.github_created_at BETWEEN $1 AND $2
-       UNION ALL
-       SELECT
-         rr.id AS review_request_id,
-         r.github_created_at AS responded_at
-       FROM review_requests_scope rr
-       JOIN reactions r ON r.subject_id = rr.pull_request_id
-       WHERE r.subject_type = 'pull_request'
-         AND r.user_id = rr.reviewer_id
-         AND r.github_created_at BETWEEN $1 AND $2
-     ),
-     valid_responses AS (
-       SELECT
-         rr.id AS review_request_id,
-         MIN(response_events.responded_at) AS responded_at
-       FROM review_requests_scope rr
-       LEFT JOIN response_events
-         ON response_events.review_request_id = rr.id
-         AND response_events.responded_at >= rr.requested_at
-         AND (rr.removed_at IS NULL OR response_events.responded_at < rr.removed_at)
-       GROUP BY rr.id
-     )
-     SELECT
-       rr.reviewer_id,
-       rr.pull_request_id,
-       rr.requested_at,
-       valid_responses.responded_at
-     FROM review_requests_scope rr
-     JOIN valid_responses ON valid_responses.review_request_id = rr.id
-     WHERE valid_responses.responded_at IS NOT NULL`,
-    params,
-  );
-
-  return result.rows;
-}
-
-type LeaderboardRow = {
-  user_id: string;
-  value: number;
-  secondary_value?: number | null;
-  additions?: number;
-  deletions?: number;
-};
-
-type PrCompletionLeaderboardRow = {
-  user_id: string;
-  value: number;
-  merged_prs: number;
-  commented_count: number;
-  changes_requested_count: number;
-};
-
-async function fetchLeaderboard(
-  metric:
-    | "issues"
-    | "reviews"
-    | "response"
-    | "comments"
-    | "prs"
-    | "prsMerged"
-    | "prsMergedBy",
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<LeaderboardRow[]> {
-  const params: unknown[] = [start, end];
-  let repoClauseIssues = "";
-  let repoClausePrs = "";
-  let repoClauseCommentsIssue = "";
-  let repoClauseCommentsPr = "";
-  let repoClauseReviews = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClauseIssues = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClausePrs = ` AND pr.repository_id = ANY($${index}::text[])`;
-    repoClauseReviews = repoClausePrs;
-    repoClauseCommentsIssue = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClauseCommentsPr = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  switch (metric) {
-    case "issues": {
-      const result = await query<LeaderboardRow>(
-        `SELECT i.author_id AS user_id, COUNT(*) AS value
-         FROM issues i
-         WHERE i.author_id IS NOT NULL AND i.github_created_at BETWEEN $1 AND $2${repoClauseIssues}
-         GROUP BY i.author_id
-         ORDER BY value DESC`,
-        params,
-      );
-      return result.rows;
-    }
-    case "prs": {
-      const result = await query<LeaderboardRow>(
-        `SELECT pr.author_id AS user_id, COUNT(*) AS value
-         FROM pull_requests pr
-         WHERE pr.author_id IS NOT NULL AND pr.github_created_at BETWEEN $1 AND $2${repoClausePrs}
-         GROUP BY pr.author_id
-         ORDER BY value DESC`,
-        params,
-      );
-      return result.rows;
-    }
-    case "prsMerged": {
-      const result = await query<LeaderboardRow>(
-        `SELECT
-           pr.author_id AS user_id,
-           COUNT(*) AS value,
-           SUM(COALESCE((pr.data ->> 'additions')::numeric, 0)) AS additions,
-           SUM(COALESCE((pr.data ->> 'deletions')::numeric, 0)) AS deletions
-         FROM pull_requests pr
-         WHERE pr.author_id IS NOT NULL AND pr.github_merged_at BETWEEN $1 AND $2${repoClausePrs}
-         GROUP BY pr.author_id
-         ORDER BY value DESC`,
-        params,
-      );
-      return result.rows;
-    }
-    case "prsMergedBy": {
-      const result = await query<LeaderboardRow>(
-        `SELECT pr.data -> 'mergedBy' ->> 'id' AS user_id, COUNT(*) AS value
-         FROM pull_requests pr
-         LEFT JOIN users u ON u.id = pr.data -> 'mergedBy' ->> 'id'
-         WHERE pr.github_merged_at BETWEEN $1 AND $2${repoClausePrs}
-           AND pr.data -> 'mergedBy' ->> 'id' IS NOT NULL
-           AND ${DEPENDABOT_FILTER}
-         GROUP BY user_id
-         ORDER BY value DESC`,
-        params,
-      );
-      return result.rows;
-    }
-    case "reviews": {
-      const result = await query<LeaderboardRow>(
-        `SELECT r.author_id AS user_id,
-                COUNT(*) FILTER (WHERE r.github_submitted_at BETWEEN $1 AND $2) AS value,
-                COUNT(DISTINCT r.pull_request_id) FILTER (WHERE r.github_submitted_at BETWEEN $1 AND $2) AS secondary_value
-         FROM reviews r
-         JOIN pull_requests pr ON pr.id = r.pull_request_id
-         WHERE r.author_id IS NOT NULL${repoClauseReviews}
-         GROUP BY r.author_id
-         ORDER BY value DESC`,
-        params,
-      );
-      return result.rows;
-    }
-    case "response": {
-      const responsePairs = await fetchReviewResponsePairs(
-        start,
-        end,
-        repositoryIds,
-      );
-      const stats = new Map<string, { sum: number; count: number }>();
-      responsePairs.forEach((row) => {
-        if (!row.reviewer_id) {
-          return;
-        }
-
-        const hours = calculateBusinessHoursBetween(
-          row.requested_at,
-          row.responded_at,
-          HOLIDAY_SET,
-        );
-        if (hours === null || !Number.isFinite(hours)) {
-          return;
-        }
-
-        const current = stats.get(row.reviewer_id) ?? { sum: 0, count: 0 };
-        current.sum += hours;
-        current.count += 1;
-        stats.set(row.reviewer_id, current);
-      });
-
-      return Array.from(stats.entries())
-        .map(([userId, { sum, count }]) => ({
-          user_id: userId,
-          value: sum / count,
-          secondary_value: count,
-        }))
-        .sort((a, b) => a.value - b.value);
-    }
-    case "comments": {
-      const result = await query<LeaderboardRow>(
-        `SELECT c.author_id AS user_id, COUNT(*) AS value
-         FROM comments c
-         LEFT JOIN issues i ON i.id = c.issue_id
-         LEFT JOIN pull_requests pr ON pr.id = c.pull_request_id
-         WHERE c.author_id IS NOT NULL AND c.github_created_at BETWEEN $1 AND $2
-           AND (
-             (c.issue_id IS NOT NULL${repoClauseCommentsIssue})
-             OR (c.pull_request_id IS NOT NULL${repoClauseCommentsPr})
-           )
-         GROUP BY c.author_id
-         ORDER BY value DESC`,
-        params,
-      );
-      return result.rows;
-    }
-    default:
-      return [];
-  }
-}
-
-async function fetchPrCompletionLeaderboard(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<PrCompletionLeaderboardRow[]> {
-  const params: unknown[] = [start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    repoClause = ` AND pr.repository_id = ANY($${params.length}::text[])`;
-  }
-
-  const result = await query<PrCompletionLeaderboardRow>(
-    `WITH merged_prs AS (
-       SELECT
-         pr.id,
-         pr.author_id,
-         pr.github_merged_at,
-         pr.repository_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_merged_at BETWEEN $1 AND $2
-         AND pr.github_merged_at IS NOT NULL
-         AND pr.author_id IS NOT NULL
-         AND ${DEPENDABOT_FILTER}
-         ${repoClause}
-     ),
-     review_totals AS (
-       SELECT
-         mp.id,
-         mp.author_id,
-         COUNT(*) FILTER (WHERE r.state = 'COMMENTED') AS commented_count,
-         COUNT(*) FILTER (WHERE r.state = 'CHANGES_REQUESTED') AS changes_requested_count
-       FROM merged_prs mp
-       LEFT JOIN reviews r
-         ON r.pull_request_id = mp.id
-        AND r.author_id IS NOT NULL
-        AND r.author_id <> mp.author_id
-        AND r.state IN ('COMMENTED', 'CHANGES_REQUESTED')
-        AND (r.github_submitted_at IS NULL OR r.github_submitted_at <= mp.github_merged_at)
-       GROUP BY mp.id, mp.author_id
-     ),
-     author_totals AS (
-       SELECT
-         author_id,
-         COUNT(*) AS merged_prs,
-         SUM(commented_count) AS commented_count,
-         SUM(changes_requested_count) AS changes_requested_count
-       FROM review_totals
-       GROUP BY author_id
-     )
-     SELECT
-       author_id AS user_id,
-       COALESCE(((commented_count + changes_requested_count)::numeric) / NULLIF(merged_prs, 0), 0) AS value,
-       merged_prs,
-       commented_count,
-       changes_requested_count
-     FROM author_totals
-     WHERE merged_prs > 0
-     ORDER BY value ASC, merged_prs DESC, author_id`,
-    params,
-  );
-
-  return result.rows.map((row) => ({
-    user_id: row.user_id,
-    value: Number(row.value ?? 0),
-    merged_prs: Number(row.merged_prs ?? 0),
-    commented_count: Number(row.commented_count ?? 0),
-    changes_requested_count: Number(row.changes_requested_count ?? 0),
-  }));
-}
-
-type IndividualIssueRow = {
-  created: number;
-  closed: number;
-  avg_resolution_hours: number | null;
-  reopened: number;
-};
-
-async function fetchIndividualIssueMetrics(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualIssueRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND i.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualIssueRow>(
-    `SELECT
-       COUNT(*) FILTER (WHERE i.github_created_at BETWEEN $2 AND $3) AS created,
-       COUNT(*) FILTER (WHERE i.github_closed_at BETWEEN $2 AND $3) AS closed,
-       AVG(EXTRACT(EPOCH FROM (i.github_closed_at - i.github_created_at)) / 3600.0)
-         FILTER (WHERE i.github_closed_at BETWEEN $2 AND $3 AND i.github_closed_at IS NOT NULL) AS avg_resolution_hours,
-       COUNT(*) FILTER (WHERE i.github_closed_at BETWEEN $2 AND $3 AND i.state = 'OPEN') AS reopened
-     FROM issues i
-     WHERE i.author_id = $1${repoClause}`,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      created: 0,
-      closed: 0,
-      avg_resolution_hours: null,
-      reopened: 0,
-    }
-  );
-}
-
-type IndividualReviewRow = {
-  reviews: number;
-  active_reviews: number;
-  avg_response_hours: number | null;
-  prs_reviewed: number;
-  active_prs_reviewed: number;
-  review_comments: number;
-};
-
-type IndividualReviewBaseRow = {
-  reviews: number;
-  prs_reviewed: number;
-  active_reviews: number;
-  active_prs_reviewed: number;
-  review_comments: number;
-};
-
-type IndividualPullRequestRow = {
-  created: number;
-  merged: number;
-};
-
-type IndividualMergedByRow = {
-  merged: number;
-};
-
-type IndividualPrCompletionRow = {
-  merged_prs: number;
-  commented_count: number;
-  changes_requested_count: number;
-};
-
-async function fetchIndividualPullRequestMetrics(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualPullRequestRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND p.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualPullRequestRow>(
-    `SELECT
-       COUNT(*) FILTER (WHERE p.github_created_at BETWEEN $2 AND $3) AS created,
-       COUNT(*) FILTER (WHERE p.github_merged_at BETWEEN $2 AND $3) AS merged
-     FROM pull_requests p
-     WHERE p.author_id = $1${repoClause}`,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      created: 0,
-      merged: 0,
-    }
-  );
-}
-
-async function fetchIndividualMergedByMetrics(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualMergedByRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualMergedByRow>(
-    `SELECT
-       COUNT(*) FILTER (WHERE pr.github_merged_at BETWEEN $2 AND $3) AS merged
-     FROM pull_requests pr
-     LEFT JOIN users u ON u.id = pr.data -> 'mergedBy' ->> 'id'
-     WHERE pr.data -> 'mergedBy' ->> 'id' = $1${repoClause}
-       AND pr.github_merged_at IS NOT NULL
-       AND ${DEPENDABOT_FILTER}`,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      merged: 0,
-    }
-  );
-}
-
-async function fetchIndividualPrCompletionMetrics(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualPrCompletionRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualPrCompletionRow>(
-    `WITH merged_prs AS (
-       SELECT
-         pr.id,
-         pr.author_id,
-         pr.github_merged_at,
-         pr.repository_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_merged_at BETWEEN $2 AND $3
-         AND pr.github_merged_at IS NOT NULL
-         AND pr.author_id = $1
-         AND ${DEPENDABOT_FILTER}
-         ${repoClause}
-     ),
-     review_totals AS (
-       SELECT
-         mp.id,
-         COUNT(*) FILTER (WHERE r.state = 'COMMENTED') AS commented_count,
-         COUNT(*) FILTER (WHERE r.state = 'CHANGES_REQUESTED') AS changes_requested_count
-       FROM merged_prs mp
-       LEFT JOIN reviews r
-         ON r.pull_request_id = mp.id
-        AND r.author_id IS NOT NULL
-        AND r.author_id <> mp.author_id
-        AND r.state IN ('COMMENTED', 'CHANGES_REQUESTED')
-        AND (r.github_submitted_at IS NULL OR r.github_submitted_at <= mp.github_merged_at)
-       GROUP BY mp.id
-     )
-     SELECT
-       COUNT(*) AS merged_prs,
-       COALESCE(SUM(review_totals.commented_count), 0) AS commented_count,
-       COALESCE(SUM(review_totals.changes_requested_count), 0) AS changes_requested_count
-     FROM merged_prs
-     LEFT JOIN review_totals ON review_totals.id = merged_prs.id`,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      merged_prs: 0,
-      commented_count: 0,
-      changes_requested_count: 0,
-    }
-  );
-}
-
-async function fetchIndividualReviewMetrics(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualReviewRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const statsResult = await query<IndividualReviewBaseRow>(
-    `WITH reviewer_reviews AS (
-       SELECT r.pull_request_id
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1
-         AND r.github_submitted_at BETWEEN $2 AND $3${repoClause}
-         AND pr.author_id <> $1
-         AND COALESCE(r.state, '') <> 'DISMISSED'
-         AND ${DEPENDABOT_FILTER}
-     ),
-     approved_reviews AS (
-       SELECT r.pull_request_id
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1
-         AND r.github_submitted_at BETWEEN $2 AND $3${repoClause}
-         AND pr.author_id <> $1
-         AND r.state = 'APPROVED'
-         AND ${DEPENDABOT_FILTER}
-     ),
-     review_comments AS (
-       SELECT COUNT(*) AS review_comments
-       FROM comments c
-       JOIN pull_requests pr ON pr.id = c.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE c.author_id = $1
-         AND c.github_created_at BETWEEN $2 AND $3${repoClause}
-         AND pr.author_id <> $1
-         AND ${DEPENDABOT_FILTER}
-     )
-     SELECT
-       (SELECT COUNT(*) FROM reviewer_reviews) AS reviews,
-       (SELECT COUNT(DISTINCT pull_request_id) FROM reviewer_reviews) AS prs_reviewed,
-       (SELECT COUNT(*) FROM approved_reviews) AS active_reviews,
-       (SELECT COUNT(DISTINCT pull_request_id) FROM approved_reviews) AS active_prs_reviewed,
-       (SELECT review_comments FROM review_comments) AS review_comments
-     `,
-    params,
-  );
-
-  const statsRow = statsResult.rows[0] ?? {
-    reviews: 0,
-    prs_reviewed: 0,
-    active_reviews: 0,
-    active_prs_reviewed: 0,
-    review_comments: 0,
-  };
-
-  const responsePairs = await fetchReviewResponsePairs(
-    start,
-    end,
-    repositoryIds,
-    personId,
-  );
-
-  const avgResponseHours = averageBusinessResponseHours(
-    responsePairs.map((row) => ({
-      requestedAt: row.requested_at,
-      respondedAt: row.responded_at,
-    })),
-    HOLIDAY_SET,
-  );
-
-  return {
-    reviews: Number(statsRow.reviews ?? 0),
-    active_reviews: Number(statsRow.active_reviews ?? 0),
-    prs_reviewed: Number(statsRow.prs_reviewed ?? 0),
-    active_prs_reviewed: Number(statsRow.active_prs_reviewed ?? 0),
-    avg_response_hours: avgResponseHours,
-    review_comments: Number(statsRow.review_comments ?? 0),
-  };
-}
-
-type IndividualCoverageRow = {
-  coverage: number | null;
-  participation: number | null;
-};
-
-async function fetchIndividualCoverageMetrics(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualCoverageRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClause = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClause = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualCoverageRow>(
-    `WITH prs_in_range AS (
-       SELECT pr.id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.github_merged_at IS NOT NULL
-         AND pr.github_merged_at BETWEEN $2 AND $3${repoClause}
-         AND ${DEPENDABOT_FILTER}
-     ),
-     reviewer_prs AS (
-       SELECT DISTINCT r.pull_request_id
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1
-         AND r.github_submitted_at BETWEEN $2 AND $3${repoClause}
-         AND pr.github_merged_at IS NOT NULL
-         AND pr.github_merged_at BETWEEN $2 AND $3
-         AND ${DEPENDABOT_FILTER}
-     ),
-     participation AS (
-       SELECT
-         pr.id,
-         COUNT(DISTINCT r.author_id) FILTER (WHERE r.github_submitted_at BETWEEN $2 AND $3) AS reviewer_count
-       FROM pull_requests pr
-       LEFT JOIN reviews r ON r.pull_request_id = pr.id
-       WHERE pr.id IN (SELECT id FROM prs_in_range)
-       GROUP BY pr.id
-     ),
-     person_participation AS (
-       SELECT
-         pr.id,
-         COUNT(DISTINCT r.author_id) FILTER (WHERE r.github_submitted_at BETWEEN $2 AND $3) AS reviewer_count
-       FROM pull_requests pr
-       JOIN reviews r ON r.pull_request_id = pr.id
-       WHERE pr.id IN (SELECT pull_request_id FROM reviewer_prs)
-       GROUP BY pr.id
-     )
-     SELECT
-       CASE WHEN (SELECT COUNT(*) FROM prs_in_range) = 0 THEN NULL
-            ELSE (SELECT COUNT(*) FROM reviewer_prs)::numeric / (SELECT COUNT(*) FROM prs_in_range)::numeric
-       END AS coverage,
-       (SELECT AVG(reviewer_count) FROM person_participation) AS participation
-     `,
-    params,
-  );
-
-  return (
-    result.rows[0] ?? {
-      coverage: null,
-      participation: null,
-    }
-  );
-}
-
-type IndividualDiscussionRow = {
-  comments: number;
-};
-
-async function fetchIndividualDiscussion(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<IndividualDiscussionRow> {
-  const params: unknown[] = [personId, start, end];
-  let repoClauseIssue = "";
-  let repoClausePr = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClauseIssue = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClausePr = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualDiscussionRow>(
-    `SELECT COUNT(*) AS comments
-     FROM comments c
-     LEFT JOIN issues i ON i.id = c.issue_id
-     LEFT JOIN pull_requests pr ON pr.id = c.pull_request_id
-     WHERE c.author_id = $1 AND c.github_created_at BETWEEN $2 AND $3
-       AND (
-         (c.issue_id IS NOT NULL${repoClauseIssue})
-         OR (c.pull_request_id IS NOT NULL${repoClausePr})
-       )`,
-    params,
-  );
-
-  return result.rows[0] ?? { comments: 0 };
-}
-
-type IndividualMonthlyRow = {
-  bucket: string;
-  issues: number;
-  reviews: number;
-};
-
-async function fetchIndividualMonthlyTrends(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-  timeZone: string,
-): Promise<MultiTrendPoint[]> {
-  const params: unknown[] = [personId, start, end];
-  let repoClauseIssues = "";
-  let repoClauseReviews = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClauseIssues = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClauseReviews = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const timezoneIndex = params.length + 1;
-  const result = await query<IndividualMonthlyRow>(
-    `WITH issue_trend AS (
-       SELECT to_char(date_trunc('month', i.github_created_at AT TIME ZONE $${timezoneIndex}), 'YYYY-MM') AS bucket,
-              COUNT(*) AS issues
-       FROM issues i
-       WHERE i.author_id = $1 AND i.github_created_at BETWEEN $2 AND $3${repoClauseIssues}
-       GROUP BY bucket
-     ),
-     review_trend AS (
-       SELECT to_char(date_trunc('month', r.github_submitted_at AT TIME ZONE $${timezoneIndex}), 'YYYY-MM') AS bucket,
-              COUNT(*) AS reviews
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1 AND r.github_submitted_at BETWEEN $2 AND $3${repoClauseReviews}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY bucket
-     ),
-     combined AS (
-       SELECT
-         COALESCE(i.bucket, r.bucket) AS bucket,
-         COALESCE(i.issues, 0) AS issues,
-         COALESCE(r.reviews, 0) AS reviews
-       FROM issue_trend i
-       FULL OUTER JOIN review_trend r ON r.bucket = i.bucket
-     )
-     SELECT bucket, issues, reviews
-     FROM combined
-     ORDER BY bucket`,
-    [...params, timeZone],
-  );
-
-  return result.rows.map((row) => ({
-    date: row.bucket,
-    values: {
-      issues: row.issues,
-      reviews: row.reviews,
-    },
-  }));
-}
-
-type IndividualRepoActivityRow = RepoDistributionRow;
-
-async function fetchIndividualRepoActivity(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<RepoDistributionRow[]> {
-  const params: unknown[] = [personId, start, end];
-  let repoFilter = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoFilter = ` AND repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<IndividualRepoActivityRow>(
-    `WITH issue_counts AS (
-       SELECT repository_id AS repo_id, COUNT(*) AS issues
-       FROM issues
-       WHERE author_id = $1 AND github_created_at BETWEEN $2 AND $3
-       GROUP BY repository_id
-     ),
-     review_counts AS (
-       SELECT pr.repository_id AS repo_id, COUNT(*) AS reviews
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1 AND r.github_submitted_at BETWEEN $2 AND $3
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id
-     ),
-     comment_counts AS (
-       SELECT repository_id AS repo_id, COUNT(*) AS comments
-       FROM (
-         SELECT i.repository_id AS repository_id
-         FROM comments c
-         JOIN issues i ON i.id = c.issue_id
-         WHERE c.author_id = $1 AND c.github_created_at BETWEEN $2 AND $3
-         UNION ALL
-         SELECT p.repository_id AS repository_id
-         FROM comments c
-         JOIN pull_requests p ON p.id = c.pull_request_id
-         WHERE c.author_id = $1 AND c.github_created_at BETWEEN $2 AND $3
-       ) combined
-       GROUP BY repository_id
-     ),
-     combined AS (
-       SELECT
-         COALESCE(ic.repo_id, rc.repo_id, cc.repo_id) AS repository_id,
-         COALESCE(ic.issues, 0) AS issues,
-         0 AS pull_requests,
-         COALESCE(rc.reviews, 0) AS reviews,
-         COALESCE(cc.comments, 0) AS comments
-       FROM issue_counts ic
-       FULL OUTER JOIN review_counts rc ON rc.repo_id = ic.repo_id
-       FULL OUTER JOIN comment_counts cc ON cc.repo_id = COALESCE(ic.repo_id, rc.repo_id)
-     )
-     SELECT *, (issues + reviews + comments) AS total_events
-     FROM combined
-     WHERE repository_id IS NOT NULL${repoFilter}
-     ORDER BY total_events DESC`,
-    params,
-  );
-
-  return result.rows;
-}
-
-async function fetchIndividualRepoComparison(
-  personId: string,
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<RepoComparisonRawRow[]> {
-  const params: unknown[] = [personId, start, end];
-  let repoFilterPr = "";
-  let repoFilterIssues = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoFilterPr = ` AND repository_id = ANY($${index}::text[])`;
-    repoFilterIssues = ` AND repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<RepoComparisonRawRow>(
-    `WITH repo_ids AS (
-       SELECT DISTINCT repository_id
-       FROM issues
-       WHERE author_id = $1
-         AND (github_created_at BETWEEN $2 AND $3 OR github_closed_at BETWEEN $2 AND $3)${repoFilterIssues}
-       UNION
-       SELECT DISTINCT pr.repository_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.author_id = $1
-         AND (pr.github_created_at BETWEEN $2 AND $3 OR pr.github_merged_at BETWEEN $2 AND $3)${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       UNION
-       SELECT DISTINCT pr.repository_id
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.data -> 'mergedBy' ->> 'id'
-       WHERE pr.data -> 'mergedBy' ->> 'id' = $1
-         AND pr.github_merged_at BETWEEN $2 AND $3${repoFilterPr}
-         AND pr.github_merged_at IS NOT NULL
-         AND ${DEPENDABOT_FILTER}
-       UNION
-       SELECT DISTINCT pr.repository_id
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE r.author_id = $1
-         AND r.github_submitted_at BETWEEN $2 AND $3${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       UNION
-       SELECT DISTINCT repository_id
-       FROM (
-         SELECT i.repository_id AS repository_id
-         FROM comments c
-         JOIN issues i ON i.id = c.issue_id
-         WHERE c.author_id = $1
-           AND c.github_created_at BETWEEN $2 AND $3${repoFilterIssues}
-         UNION ALL
-         SELECT p.repository_id AS repository_id
-         FROM comments c
-         JOIN pull_requests p ON p.id = c.pull_request_id
-         LEFT JOIN users u ON u.id = p.author_id
-         WHERE c.author_id = $1
-           AND c.github_created_at BETWEEN $2 AND $3${repoFilterPr}
-           AND ${DEPENDABOT_FILTER}
-       ) AS comment_repo_ids
-     ),
-     issue_created_counts AS (
-       SELECT repository_id, COUNT(*) AS issues_created
-       FROM issues
-       WHERE author_id = $1
-         AND github_created_at BETWEEN $2 AND $3${repoFilterIssues}
-       GROUP BY repository_id
-     ),
-     issue_counts AS (
-       SELECT repository_id, COUNT(*) AS issues_resolved
-       FROM issues
-       WHERE author_id = $1
-         AND github_closed_at BETWEEN $2 AND $3${repoFilterIssues}
-       GROUP BY repository_id
-     ),
-     pr_created_counts AS (
-       SELECT pr.repository_id, COUNT(*) AS prs_created
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.author_id = $1
-         AND pr.github_created_at BETWEEN $2 AND $3${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id
-     ),
-     pr_counts AS (
-       SELECT pr.repository_id, COUNT(*) AS prs_merged
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       WHERE pr.author_id = $1
-         AND pr.github_merged_at BETWEEN $2 AND $3${repoFilterPr}
-         AND pr.github_merged_at IS NOT NULL
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id
-     ),
-     pr_merged_by_counts AS (
-       SELECT pr.repository_id, COUNT(*) AS prs_merged_by
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.data -> 'mergedBy' ->> 'id'
-       WHERE pr.data -> 'mergedBy' ->> 'id' = $1
-         AND pr.github_merged_at BETWEEN $2 AND $3${repoFilterPr}
-         AND pr.github_merged_at IS NOT NULL
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id
-     ),
-    review_counts AS (
-      SELECT
-        pr.repository_id,
-        COUNT(*) AS reviews,
-        COUNT(*) FILTER (WHERE r.state = 'APPROVED') AS active_reviews
-      FROM reviews r
-      JOIN pull_requests pr ON pr.id = r.pull_request_id
-      LEFT JOIN users u ON u.id = pr.author_id
-      WHERE r.author_id = $1
-        AND r.github_submitted_at BETWEEN $2 AND $3${repoFilterPr}
-        AND COALESCE(r.state, '') <> 'DISMISSED'
-        AND ${DEPENDABOT_FILTER}
-      GROUP BY pr.repository_id
-    ),
-     comment_counts AS (
-       SELECT repository_id, COUNT(*) AS comments
-       FROM (
-         SELECT i.repository_id AS repository_id
-         FROM comments c
-         JOIN issues i ON i.id = c.issue_id
-         WHERE c.author_id = $1
-           AND c.github_created_at BETWEEN $2 AND $3${repoFilterIssues}
-         UNION ALL
-         SELECT p.repository_id AS repository_id
-         FROM comments c
-         JOIN pull_requests p ON p.id = c.pull_request_id
-         LEFT JOIN users u ON u.id = p.author_id
-         WHERE c.author_id = $1
-           AND c.github_created_at BETWEEN $2 AND $3${repoFilterPr}
-           AND ${DEPENDABOT_FILTER}
-       ) AS combined
-       GROUP BY repository_id
-     ),
-     first_review_times AS (
-       SELECT
-         pr.repository_id,
-         pr.github_created_at,
-         MIN(r.github_submitted_at) AS first_review_at
-       FROM pull_requests pr
-       LEFT JOIN users u ON u.id = pr.author_id
-       LEFT JOIN reviews r ON r.pull_request_id = pr.id
-       WHERE pr.author_id = $1
-         AND pr.github_created_at BETWEEN $2 AND $3${repoFilterPr}
-         AND ${DEPENDABOT_FILTER}
-       GROUP BY pr.repository_id, pr.id, pr.github_created_at
-     ),
-     first_reviews AS (
-       SELECT
-         repository_id,
-         AVG(EXTRACT(EPOCH FROM (first_review_at - github_created_at)) / 3600.0) AS avg_first_review_hours
-       FROM first_review_times
-       WHERE first_review_at IS NOT NULL
-       GROUP BY repository_id
-     )
-     SELECT
-       repo_ids.repository_id,
-       COALESCE(issue_created_counts.issues_created, 0) AS issues_created,
-       COALESCE(issue_counts.issues_resolved, 0) AS issues_resolved,
-       COALESCE(pr_created_counts.prs_created, 0) AS prs_created,
-       COALESCE(pr_counts.prs_merged, 0) AS prs_merged,
-       COALESCE(pr_merged_by_counts.prs_merged_by, 0) AS prs_merged_by,
-       COALESCE(review_counts.reviews, 0) AS reviews,
-       COALESCE(review_counts.active_reviews, 0) AS active_reviews,
-       COALESCE(comment_counts.comments, 0) AS comments,
-       first_reviews.avg_first_review_hours
-     FROM repo_ids
-     LEFT JOIN issue_created_counts ON issue_created_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN issue_counts ON issue_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN pr_created_counts ON pr_created_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN pr_counts ON pr_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN pr_merged_by_counts ON pr_merged_by_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN review_counts ON review_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN comment_counts ON comment_counts.repository_id = repo_ids.repository_id
-     LEFT JOIN first_reviews ON first_reviews.repository_id = repo_ids.repository_id`,
-    params,
-  );
-
-  return result.rows;
-}
-
-async function fetchActiveContributors(
-  start: string,
-  end: string,
-  repositoryIds: string[] | undefined,
-): Promise<string[]> {
-  const params: unknown[] = [start, end];
-  let repoClauseIssues = "";
-  let repoClausePrs = "";
-  let repoClauseReviews = "";
-  let repoClauseCommentsIssue = "";
-  let repoClauseCommentsPr = "";
-  if (repositoryIds?.length) {
-    params.push(repositoryIds);
-    const index = params.length;
-    repoClauseIssues = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClausePrs = ` AND pr.repository_id = ANY($${index}::text[])`;
-    repoClauseReviews = repoClausePrs;
-    repoClauseCommentsIssue = ` AND i.repository_id = ANY($${index}::text[])`;
-    repoClauseCommentsPr = ` AND pr.repository_id = ANY($${index}::text[])`;
-  }
-
-  const result = await query<{ user_id: string }>(
-    `WITH issue_authors AS (
-       SELECT DISTINCT i.author_id AS user_id
-       FROM issues i
-       WHERE i.author_id IS NOT NULL AND i.github_created_at BETWEEN $1 AND $2${repoClauseIssues}
-     ),
-     review_authors AS (
-       SELECT DISTINCT r.author_id AS user_id
-       FROM reviews r
-       JOIN pull_requests pr ON pr.id = r.pull_request_id
-       WHERE r.author_id IS NOT NULL AND r.github_submitted_at BETWEEN $1 AND $2${repoClauseReviews}
-     ),
-     comment_authors AS (
-       SELECT DISTINCT c.author_id AS user_id
-       FROM comments c
-       LEFT JOIN issues i ON i.id = c.issue_id
-       LEFT JOIN pull_requests pr ON pr.id = c.pull_request_id
-       WHERE c.author_id IS NOT NULL AND c.github_created_at BETWEEN $1 AND $2
-         AND (
-           (c.issue_id IS NOT NULL${repoClauseCommentsIssue})
-           OR (c.pull_request_id IS NOT NULL${repoClauseCommentsPr})
-         )
-     )
-     SELECT DISTINCT user_id
-     FROM (
-       SELECT user_id FROM issue_authors
-       UNION
-       SELECT user_id FROM review_authors
-       UNION
-       SELECT user_id FROM comment_authors
-     ) combined
-     WHERE user_id IS NOT NULL`,
-    params,
-  );
-
-  return result.rows.map((row) => row.user_id);
-}
-
-async function resolveProfiles(
-  repositoryIds: string[],
-  userIds: string[],
-): Promise<{ repositories: RepositoryProfile[]; users: UserProfile[] }> {
-  const [repositories, users] = await Promise.all([
-    repositoryIds.length
-      ? getRepositoryProfiles(repositoryIds)
-      : Promise.resolve([]),
-    userIds.length ? getUserProfiles(userIds) : Promise.resolve([]),
-  ]);
-
-  return { repositories, users };
-}
 
 function mapRepoDistribution(
   rows: RepoDistributionRow[],
@@ -2126,7 +197,16 @@ function mapPrCompletionLeaderboard(
 
       return {
         user,
-        value: Number(row.value ?? 0),
+        value: (() => {
+          const merged = Number(row.merged_prs ?? 0);
+          if (merged === 0) {
+            return 0;
+          }
+          const interactions =
+            Number(row.commented_count ?? 0) +
+            Number(row.changes_requested_count ?? 0);
+          return interactions / merged;
+        })(),
         secondaryValue: Number(row.merged_prs ?? 0),
         details: [
           {
